feat: migrate owner commands to messages
