+++ src/modules/naomi/naomiViewCommands.ts
@@ -1,0 +1,55 @@
+import { REST } from "@discordjs/rest";
+import { Routes } from "discord-api-types/v9";
+import { Message, MessageEmbed } from "discord.js";
+
+import { BeccaLyria } from "../../interfaces/BeccaLyria";
+import { CommandData } from "../../interfaces/commands/CommandData";
+import { beccaErrorHandler } from "../../utils/beccaErrorHandler";
+
+/**
+ * Owner-only module to view the current list of commands.
+ *
+ * @param {BeccaLyria} Becca Becca's discord instance.
+ * @param {Message} message The message payload from Discord.
+ */
+export const naomiViewCommands = async (
+  Becca: BeccaLyria,
+  message: Message
+) => {
+  try {
+    const rest = new REST({ version: "9" }).setToken(Becca.configs.token);
+
+    const commands: CommandData[] = (await rest.get(
+      Routes.applicationCommands(Becca.configs.id)
+    )) as CommandData[];
+
+    if (!commands.length) {
+      await message.reply(
+        "It seems I do not have any spells prepared at this time."
+      );
+      return;
+    }
+
+    const embed = new MessageEmbed();
+    embed.setTitle("Available Spells");
+    embed.setDescription("These are the spells I currently have prepared.");
+
+    for (const command of commands) {
+      embed.addField(
+        command.name,
+        command.options?.map((opt) => opt.name).join(", ") ||
+          "This spell has no options."
+      );
+    }
+
+    await message.reply({ embeds: [embed] });
+  } catch (err) {
+    await beccaErrorHandler(
+      Becca,
+      "view slash command",
+      err,
+      message.guild?.name,
+      message
+    );
+  }
+};
